requirements.txt
slither-analyzer
anthropic
PyGithub
fastapi
uvicorn
schedule
requests
python-dotenv
